backend/requirements.txt
Django>=5.2.7
python-dotenv>=1.0.0
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
django-filter>=23.0
